Another pass on kformat_bytes
I realized I could _probably_ do it as a func, since it's only calling
funcs ... and while I lose the ability to do faux local variables by
sticking things into a record ... I can decompose it with const and
other func calls.

AND - magically up in the higher end of the EB range, it works again! So
clearly there was a bug somewhere in the pipeline version of this, in
the `super` internals that might(?) be worth exploring? Probably with
the division calls?