Big install push
- add sk, skdoc, skgrok bin - shell scripts
- change install to pull from repo and get everything into XDG locations

I'll need to iterate on this a bit, and I'll just do this on `main`
since it's early days -- probably in the future I'll need to be able to
iterate the installer on a branch, but I can solve that later.

I also definitely need to move to packaging everything up so the install
script can just pull one file from the repo and then divvy it all up.
But I want to do some testing here first, thus the hardcoded list of doc
files it pulls for right now.