spq-doc.sh: inline jsup, not comments to parse
I was staring at an open file, and while being able to parse the strings
with super was nice, even better to just inline the skdoc record into a
related skdoc_ func, right? So much simpler to write, parse, etc.

I'm also starting the prefix rename from `k` to `sk_` ... cuz that's
stewed in my brain a few days now and I'm cool with it.