skops: less needs LESSCHARSET=utf-8
While I don't like the hacked output compressing all example output
lines into a single line with a carriage return character, I don't have
any quick ideas for something better, and we can iterate on it later.
Having skops at all is pretty useful.

The goal is a quick reference, something compact and quick on a
terminal, not the ultimate in docs - the website is much better for all
of that.